% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbindex.R
\name{comorbindex}
\alias{comorbindex}
\title{Compute the AHRQ comorbidity index for a dataset.}
\usage{
comorbindex(df, indextype = "mort")
}
\arguments{
\item{df}{A data frame of comorbidity indicators (1 if the comorbidity is
present, 0 if not).  There must be a column for each of the comorbidities
in the index.}

\item{indextype}{Which index to calculate, see details}
}
\value{
A vector of comorbidity index values
}
\description{
Given a table of comorbidity indicator variables, compute the AHRQ index
for risk of in-hospital mortality.
}
\details{
AHRQ has produced risk indices for two outcomes:  in-hospital mortality
('mort') and all-cause readmission ('readm').  In the future we may add
some indices for USNWR outcomes 30-day survival and discharge to home.
Adding these isn't entirely straightforward because they are calculated for
individual specialties and conditions.  The default is 'mort'.
}
