% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{integer_logic}
\alias{integer_logic}
\alias{intor}
\alias{pintand}
\title{Logical reductions, with integer return}
\usage{
intor(x)

pintand(x, y)
}
\arguments{
\item{x, y}{Vectors of numeric values to be interpreted as logical values}
}
\description{
Perform a logical reduction on a numeric values and return the result as
an integer.
}
\details{
So far we've only implemented the combinations (vector/parallel x and/or) that
are actually used in package functions.
}
\section{Functions}{
\itemize{
\item \code{intor}: Vector integer or

\item \code{pintand}: Parallel integer and
}}

\keyword{internal}
