% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{keyfuns}
\alias{keyfuns}
\alias{mergekeys}
\alias{splitkeys}
\title{Convenience functions for merging and separating keys}
\usage{
mergekeys(tbl, idcols, synthid = "id", drop = TRUE, sepchar = intToUtf8(31))

splitkeys(tbl, idcols, synthid = "id", drop = TRUE, sepchar = intToUtf8(31))
}
\arguments{
\item{tbl}{Input table}

\item{idcols}{Names of the columns with the separated keys}

\item{synthid}{Name of the column with the synthetic key}

\item{drop}{If \code{TRUE}, remove the input column(s) from the output}

\item{sepchar}{Character to use to separate elements in the synthetic key
(default is ASCII Unit Separator (\code{^_})}
}
\description{
Convenience functions for merging and separating keys
}
\section{Functions}{
\itemize{
\item \code{mergekeys()}: Merge two or more columns into a synthetic key

\item \code{splitkeys()}: Separate synthetic key into two or more columns

}}
