% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poa_functions.R
\name{poa_functions}
\alias{poa_functions}
\alias{poa_cms}
\alias{poa_cms2}
\alias{poa_natural}
\title{Functions for decoding POA flags}
\usage{
poa_cms(poa)

poa_cms2(poa)

poa_natural(poa)
}
\arguments{
\item{poa}{Vector of POA values to be decoded.}
}
\value{
Integer vector of 1 (true), 0 (false), and NA (other) values.
}
\description{
There are a variety of conventions for representing the true/false/other
values that come up in the POA fields.  They differ somewhat in what they
use for true and false values, but the types of "other" allowed (if any) are
all over the map.  These functions decode several conventions in common use.
}
\details{
In some cases the distinction between the different types of "other" results is
fuzzy.  For example, there is a case to be made that any entry other than a
clear "true" should be treated as "false", since the condition was not in fact
reported as POA.  These functions provide some flexibility in this treatment.

\itemize{
\item \code{poa_cms} treats Y as true (\emph{uppercase only}), and any other result as false.
}

\itemize{
\item \code{poa_cms2} treats Y as true, N as false (\emph{uppercase only}), and anything else as "other"
}

\itemize{
\item \code{poa_natural} looks at the starting character and treats Y, T, and 1
as true; N, F, and 0 as false, and everything else as "other".
}
}
\section{Functions}{
\itemize{
\item \code{poa_cms()}: CMS POA convention, strict interpretation

\item \code{poa_cms2()}: CMS POA convention

\item \code{poa_natural()}: A default convention that reflects common usages.

}}
